fix playerTypeQuestionnaireScreen ( multiple clicks bug + ensuring integrity). add leadboard design for first player if highlighted

+++ app/src/main/java/com/example/habitspark/data/models/metaUserGroupsModel.kt
@@ -1,8 +1,8 @@
 package com.example.habitspark.data.models
 
 data class metaUserGroupsModel (
-    val totalAssgined: Int = 0,
-    val A_ALL_Count: Int = 0,
-    val B_GATED_count: Int = 0
+    val totalAssgined: Long = 0,
+    val A_ALL_Count: Long = 0,
+    val B_GATED_count: Long = 0
 
 )
